+++ CHANGELOG.md
@@ -48,6 +48,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - admin_article_variant_bottom_extended
 - New methods:
   - `OxidEsales\EshopCommunity\Core\InputValidator::addValidationError`
+  - `OxidEsales\EshopCommunity\Application\Controller\Admin\ActionsMain::checkAccessToEditAction()`
+  - `OxidEsales\EshopCommunity\Application\Controller\Admin\AdminController::isNewEditObject()`
+  - `OxidEsales\EshopCommunity\Application\Model\Actions::isDefault()`
 
 ### Changed
 - In voucher series generation, if Coupon Number radio button checked, the number is marked as Required now. [PR-676](https://github.com/OXID-eSales/oxideshop_ce/pull/676)